+++ Source/FicsitWiremod/Public/Behaviour/Gates/Arrays/ArrayLength.h
@@ -12,7 +12,7 @@ class FICSITWIREMOD_API AArrayLength : public AFGWiremodBuildable
 	GENERATED_BODY()
     
 public:
-	virtual void Process_Implementation(float DeltaTime) override
+	virtual void Process_Implementation(double DeltaTime) override
 	{
 		Out = GetArrayLength(GetConnection(0));
 	}
