+++ Source/FicsitWiremod/Public/Behaviour/Gates/Logic/Timer.h
@@ -12,7 +12,7 @@ class FICSITWIREMOD_API ATimer : public AFGWiremodBuildable
 	GENERATED_BODY()
 
 public:
-	virtual void Process_Implementation(float DeltaTime) override
+	virtual void Process_Implementation(double DeltaTime) override
 	{
 		Out = false;
 
@@ -47,5 +47,5 @@ public:
 	bool Out;
 
 	UPROPERTY(Replicated, SaveGame)
-	float CurrentTime;
+	double CurrentTime;
 };
