+++ Source/FicsitWiremod/Public/Behaviour/Gates/Numbers/InverseNumber.h
@@ -12,7 +12,7 @@ class FICSITWIREMOD_API AInverseNumber : public AFGWiremodBuildable
 	GENERATED_BODY()
     
 public:
-	virtual void Process_Implementation(float DeltaTime) override
+	virtual void Process_Implementation(double DeltaTime) override
 	{
 		Out = 1.f / GetConnection(0).GetFloat();;
 	}
@@ -25,5 +25,5 @@ public:
 	}
 	
 	UPROPERTY(Replicated, VisibleInstanceOnly)
-	float Out;
+	double Out;
 };
